Break down vendor bundles using code-splitting
