+++ package.json
@@ -29,7 +29,6 @@
     "date-fns": "^1.30.1",
     "jquery": "^3.4.1",
     "lodash": "^4.17.11",
-    "object-hash": "^1.3.1",
     "v-tooltip": "^2.0.2",
     "vue": "^2.6.10",
     "vue-carousel": "^0.18.0",
@@ -55,6 +54,7 @@
     "require-from-string": "^2.0.2",
     "sass": "^1.18.0",
     "sass-loader": "^7.1.0",
-    "vue-template-compiler": "^2.6.10"
+    "vue-template-compiler": "^2.6.10",
+    "webpack-bundle-analyzer": "^3.3.2"
   }
 }
