+++ src/components/Now/ProductivityCard/index.vue
@@ -21,31 +21,14 @@
 
 <script>
 import * as yup from "yup";
-import { Doughnut, mixins } from "vue-chartjs";
-import { kebabCase, isEmpty } from "lodash";
+import isEmpty from "lodash/isEmpty";
+import kebabCase from "lodash/kebabCase";
 
-import Card from "./Card";
+import Card from "../Card";
 import { segment } from "@/schemas/productivity";
 
-const Chart = {
-  extends: Doughnut,
-  mixins: [mixins.reactiveProp],
-  data: () => ({
-    options: {
-      legend: { display: false },
-      responsive: true,
-      cutoutPercentage: 70,
-      tooltips: {
-        callbacks: {
-          label: (item, data) => " " + data.labels[item.index],
-        },
-      },
-    },
-  }),
-  mounted() {
-    this.renderChart(this.chartData, this.options);
-  },
-};
+const Chart = () =>
+  import(/* webpackChunkName: "productivity-chart" */ "./Chart");
 
 export default {
   props: {
